rename hillshade namespace -> hillshader
