+++ code/src/hillshader/CMakeLists.txt
@@ -1,11 +1,11 @@
 set(HILLSHADER_FILES
-    "${CMAKE_CURRENT_SOURCE_DIR}/cpp/hillshade/main.cpp"
-    "${CMAKE_CURRENT_SOURCE_DIR}/cpp/hillshade/application.cpp"
-    "${CMAKE_CURRENT_SOURCE_DIR}/cpp/hillshade/terrain.cpp"
-    "${CMAKE_CURRENT_SOURCE_DIR}/cpp/hillshade/mesh.cpp"
-    "${CMAKE_CURRENT_SOURCE_DIR}/include/private/hillshade/application.hpp"
-    "${CMAKE_CURRENT_SOURCE_DIR}/include/private/hillshade/terrain.hpp"
-    "${CMAKE_CURRENT_SOURCE_DIR}/include/private/hillshade/mesh.hpp"
+    "${CMAKE_CURRENT_SOURCE_DIR}/cpp/hillshader/main.cpp"
+    "${CMAKE_CURRENT_SOURCE_DIR}/cpp/hillshader/application.cpp"
+    "${CMAKE_CURRENT_SOURCE_DIR}/cpp/hillshader/terrain.cpp"
+    "${CMAKE_CURRENT_SOURCE_DIR}/cpp/hillshader/mesh.cpp"
+    "${CMAKE_CURRENT_SOURCE_DIR}/include/private/hillshader/application.hpp"
+    "${CMAKE_CURRENT_SOURCE_DIR}/include/private/hillshader/terrain.hpp"
+    "${CMAKE_CURRENT_SOURCE_DIR}/include/private/hillshader/mesh.hpp"
 )
 
 add_executable(hillshader WIN32 ${HILLSHADER_FILES})
