+++ code/src/hillshader/cpp/hillshader/mesh.cpp
