+++ code/src/hillshader/cpp/hillshader/terrain.cpp
@@ -8,7 +8,7 @@
 #define STB_IMAGE_IMPLEMENTATION
 #include <stb_image.h>
 
-namespace hillshade
+namespace hillshader
 {
 
     terrain::terrain(std::filesystem::path const& path) :
