+++ code/src/hillshader/include/private/hillshader/mesh.hpp
